Improved the memory usage.
- Properties can now be reset in order to free memory without triggering change events.
- Workflows within grapevine can be made atomic. If a workflow is atomic, input and output properties will be reset, once a module has been updated. This ensures that all memory can be freed once its no longer needed by future modules.
- Training modules of ConvRBMs are more memory efficient. No copy of the model is created before training. Also, training data and data from a previous module will be freed if it's not needed by other modules.

+++ gapputils/gml.convrbm4d/Model.h
@@ -23,7 +23,7 @@ struct ModelChecker { ModelChecker(); };
 class Model : public capputils::reflection::ReflectableClass {
 public:
   const static unsigned dimCount = 4;
-  typedef double value_t;
+  typedef float value_t;
   typedef tbblas::tensor<value_t, dimCount> tensor_t;
   typedef tensor_t::dim_t dim_t;
